app-health-back/src/@api/app-health/application-authorization/resolvers/app-health-delete-application-authorizations.resolver.ts
import { AppHealthDeleteApplicationAuthorizationsHandler } from '@api/app-health/application-authorization';
import { AppHealthApplicationAuthorization } from '@api/graphql';
import { Auditing, AuditingMeta, QueryStatement, Timezone } from '@aurorajs.dev/core';
import { Args, Mutation, Resolver } from '@nestjs/graphql';

@Resolver()
export class AppHealthDeleteApplicationAuthorizationsResolver
{
    constructor(
        private readonly handler: AppHealthDeleteApplicationAuthorizationsHandler,
    ) {}

    @Mutation('appHealthDeleteApplicationAuthorizations')
    async main(
        @Args('query') queryStatement?: QueryStatement,
        @Args('constraint') constraint?: QueryStatement,
        @Timezone() timezone?: string,
    ): Promise<AppHealthApplicationAuthorization[]>
    {
        return await this.handler.main(
            queryStatement,
            constraint,
            timezone,
        );
    }
}
